Remove cell numbers from dge output file names
It's not planned to extract dge for varying cell numbers, so cell numbers in file names are not required to distinguish files. They make the workflow overly complicated, because these files require special input functions in downstream analyses and "all rules". Furthermore, changing the cell number and re-running the workflow wouldn't delete the old files and lead to cluttered data directories.